pyproject.toml
[tool.poetry]
name = "moderatsiya-kartochek"
version = "0.1.0"
description = "Классификация изображений на предмет курения"
authors = ["Georgii Chaikin <chaykin.georgiy217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
openvino = "^2024.3.0"
pillow = "^10.4.0"
transformers = "^4.44.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[tool.poetry.group.train.dependencies]
jupyterlab = "^4.2.5"
ipywidgets = "^8.1.5"
torch = { version = "^2.4.0+cu118", source = "torch_gpu" }
torchvision = { version = "^0.19.0+cu118", source = "torch_gpu" }
transformers = { extras = ["torch"], version = "^4.44.2" }
datasets = "^2.21.0"
evaluate = "^0.4.2"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"

[[tool.poetry.source]]
name = "torch_gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.poetry.scripts]
make_submission = "moderatsiya_kartochek.main:make_submission"
